+++ lib/urepo/plugs/exporter.ex
@@ -1,0 +1,5 @@
+defmodule Urepo.Plugs.Exporter do
+  @moduledoc false
+
+  use Prometheus.PlugExporter
+end