+++ lib/urepo/plugs/instrumenter.ex
@@ -1,0 +1,5 @@
+defmodule Urepo.Plugs.Instrumenter do
+  @moduledoc false
+
+  use Prometheus.PlugPipelineInstrumenter
+end